added add productvariation api

+++ repositories/ProductVariationRepository.java
@@ -3,5 +3,5 @@ package com.Bootcamp.Project.Application.repositories;
 import com.Bootcamp.Project.Application.entities.ProductVariation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductVariationRepository extends CrudRepository<ProductVariation,Integer> {
+public interface ProductVariationRepository extends CrudRepository<ProductVariation,Long> {
 }
